+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -11,6 +11,11 @@ CHANGELOG
    changed from `Serializer` to `PhpSerializer` inside `AmqpReceiver`,
    `AmqpSender`, `AmqpTransport` and `AmqpTransportFactory`.
 
+ * Added `TransportException` to mark an exception transport-related
+
+ * [BC BREAK] If listening to exceptions while using `AmqpSender` or `AmqpReceiver`, `\AMQPException` is
+   no longer thrown in favor of `TransportException`.
+
 4.2.0
 -----